+++ src/Objects/Line2D.cpp
@@ -15,5 +15,5 @@ Line2D::Line2D(int x1, int y1, int x2, int y2) {
 }
 
 Point2D Line2D::middlePoint() const {
-    return Point2D((x1 + x2) / 2, (y1 + y2) / 2);
+    return Point2D((float) (x1 + x2) / 2, (float) (y1 + y2) / 2);
 }
